+++ src/etphoneshome/objects/Hitbox.java
@@ -2,21 +2,21 @@ package etphoneshome.objects;
 
 public class Hitbox {
     private Location topLeftCorner; //Change l to L for real code
-    private double height, width;
+    private int height, width;
 
     //Constructor
-    public Hitbox(Location m, double h, double w) {
+    public Hitbox(Location m, int h, int w) {
         topLeftCorner = m;
         height = h;
         width = w;
     }
 
     //Getter
-    public double getWidth() {
+    public int getWidth() {
         return width;
     }
 
-    public double getHeight() {
+    public int getHeight() {
         return height;
     }
 
@@ -32,15 +32,15 @@ public class Hitbox {
     //Collision
     public boolean areColliding(Hitbox otherObject) {
 
-        double leftSideX = topLeftCorner.getXcord();
-        double rightSideX = topLeftCorner.getXcord() + width;
-        double otherRightSideX = otherObject.topLeftCorner.getXcord() + otherObject.getWidth(); //Make a getLocation???
-        double otherLeftSideX = otherObject.topLeftCorner.getXcord();
+        int leftSideX = topLeftCorner.getXcord();
+        int rightSideX = topLeftCorner.getXcord() + width;
+        int otherRightSideX = otherObject.topLeftCorner.getXcord() + otherObject.getWidth(); //Make a getLocation???
+        int otherLeftSideX = otherObject.topLeftCorner.getXcord();
 
-        double topSideY = topLeftCorner.getYcord();
-        double bottomSideY = topLeftCorner.getYcord() + height;
-        double otherTopSideY = otherObject.topLeftCorner.getYcord();
-        double otherBottomSideY = otherObject.topLeftCorner.getYcord() + otherObject.getHeight();
+        int topSideY = topLeftCorner.getYcord();
+        int bottomSideY = topLeftCorner.getYcord() + height;
+        int otherTopSideY = otherObject.topLeftCorner.getYcord();
+        int otherBottomSideY = otherObject.topLeftCorner.getYcord() + otherObject.getHeight();
 
         if (rightSideX < otherLeftSideX) {
             return false;
